seblu-push: Remove PKGBUILD sourcing
When PKGBUILD has pkgver function, sourcing it is not enough to get the package
version (used to guess the package name to push). In that case, we need to
dowload the sources and execute code to compute this version. This is heavy
and has security implication beyond push fonction.

We could have a partial PKGBUILD sourcing support by only working with the few
which doesn't have pkgver function. But this would lead to difficult to
diagnosis and the feature not worth it.